code/pyproject.toml
[tool.poetry]
name = "thesis"
version = "0.1.0"
description = ""
authors = ["Harish Rajagopal <harish.rajagopals@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
torch = "^1.11.0"
tensorboard = "^2.9.0"
pytorch-lightning = "^1.6.3"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"
ipykernel = "^6.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.pydocstyle]
convention = "google"

[tool.mypy]
ignore_missing_imports = true
